Refactor: routes/helper moved to routes/helper/index
